WIP: New package `document-types-ts` to decouple doc manipulation.
Decouple it from the frontend, that is.
