+++ .oxfmtrc.json
@@ -11,7 +11,7 @@
         "customGroups": [
             {
                 "groupName": "workspace",
-                "elementNamePattern": ["catcolab-ui-components", "catcolab-api", "catlog-wasm"],
+                "elementNamePattern": ["catcolab-ui-components", "catcolab-api", "catlog-wasm", "document-types"],
             },
         ],
         "groups": [
