--- packages/document-types-rs/package.json
@@ -1,13 +1,0 @@
-{
-    "name": "document-types",
-    "private": true,
-    "license": "MIT",
-    "version": "0.1.0",
-    "type": "module",
-    "scripts": {
-        "build:node": "wasm-pack build ./. --target nodejs -d ./dist/pkg-node"
-    },
-    "devDependencies": {
-        "wasm-pack": "^0.14.0"
-    }
-}
