+++ packages/document-types-ts/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "document-types",
+    "version": "0.1.0",
+    "private": true,
+    "license": "MIT",
+    "type": "module",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "scripts": {
+        "build:node": "wasm-pack build ../document-types-rs --target nodejs -d ../document-types-ts/dist/pkg-node",
+        "build": "pnpm run build:node && tsc",
+        "format": "oxfmt .",
+        "lint": "oxlint --fix && oxfmt .",
+        "check": "pnpm run lint && tsc",
+        "ci": "oxlint --deny-warnings && oxfmt --check ."
+    },
+    "dependencies": {
+        "tiny-invariant": "^1.3.3",
+        "uuid": "^13.0.0"
+    },
+    "devDependencies": {
+        "@types/uuid": "^10.0.0",
+        "typescript": "^5.2.2",
+        "wasm-pack": "^0.14.0"
+    }
+}
