+++ packages/document-types-ts/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./types";
+
+export * as Nb from "./notebook";
