+++ packages/frontend/package.json
@@ -5,10 +5,11 @@
     "license": "MIT",
     "type": "module",
     "scripts": {
+        "build:doc-types": "pnpm --prefix ../document-types-ts run build",
         "build:wasm": "pnpm --prefix ../catlog-wasm run build:browser",
         "build:bindings": "cargo run -p backend -- generate-bindings",
         "build:tcm": "tcm -p 'src/**/*.module.css' .",
-        "build:deps": "pnpm run build:wasm && pnpm run build:bindings && pnpm run build:tcm",
+        "build:deps": "pnpm run build:doc-types && pnpm run build:wasm && pnpm run build:bindings && pnpm run build:tcm",
         "build": "tsc -b && vite build --sourcemap true",
         "build:staging": "tsc -b && vite build --sourcemap true --mode staging",
         "preview": "vite preview",
@@ -49,6 +50,7 @@
         "catcolab-api": "link:../backend/pkg",
         "catcolab-ui-components": "link:../ui-components",
         "catlog-wasm": "link:../catlog-wasm/dist/pkg-browser",
+        "document-types": "link:../document-types-ts",
         "echarts": "^5.5.1",
         "echarts-solid": "^0.2.0",
         "elkjs": "^0.11.0",
