+++ packages/frontend/src/notebook/index.ts
@@ -1,3 +1,2 @@
 export * from "./notebook_cell";
 export * from "./notebook_editor";
-export * from "./types";
