+++ packages/frontend/src/notebook/notebook_cell.tsx
@@ -17,7 +17,7 @@ import type { EditorView } from "prosemirror-view";
 import { createEffect, createSignal, type JSX, onCleanup, Show } from "solid-js";
 
 import { type Completion, Completions, IconButton, InlineInput } from "catcolab-ui-components";
-import type { Uuid } from "catlog-wasm";
+import type { Uuid } from "document-types";
 import { RichTextEditor } from "../components";
 
 import "./notebook_cell.css";
