+++ pnpm-workspace.yaml
@@ -4,6 +4,6 @@ packages:
   - "dev-docs"
   - "packages/backend/pkg"
   - "packages/frontend"
-  - "packages/document-types-rs"
+  - "packages/document-types-ts"
   - "packages/ui-components"
   - "packages/gaios"
